updated create_hf_model so that it uses bf16 every train

+++ applications/DeepSpeed-Chat/dschat/utils/model/model_utils.py
@@ -101,10 +101,12 @@ def create_hf_model(model_class,
         # the weight loading is handled by create critic model
         model = model_class.from_config(model_config)
     else:
+        # 2/21 - forcing bf16 data type
         model = model_class.from_pretrained(
             model_name_or_path,
             from_tf=bool(".ckpt" in model_name_or_path),
-            config=model_config)
+            config=model_config,
+            torch_dtype=torch.bfloat16)
 
     model.config.end_token_id = tokenizer.eos_token_id
     model.config.pad_token_id = model.config.eos_token_id
